\name{getOutput}
\alias{getOutput}
\title{
get cleaner output from POLYPHEMUS results
}
\description{
data postprocessing function, return a list of table with 3 columns (ID, ratio, approx), used by default in getResult function.
}
\usage{
getOutput(inputTable = NULL, filename = NULL, norm, stretch, tssL, chr, outputdir, denoString, time, toPlot = NULL)
}
\arguments{
  \item{inputTable}{
	\code{data.frame}, polyphemus output, if NULL filename is used 	.
}
  \item{filename}{
	\code{String}, the path and name of polyphemus output if NULL the inputTable is used.
}
  \item{norm}{
	\code{String}, normalization method
}
  \item{stretch}{
   \code{numeric}, if a length value is specified, the ChIP seq ratio profiles will be "stretched"  or compacted to reach the length value. if NULL is specified the median length of the overall profiles is used.
}
  \item{tssL}{
	\code{numeric}, estimated length for the TSS region. The TSS part of the coding region will be not "stretched".
}
  \item{chr}{
	\code{numeric}, the chromosome index.
}
  \item{outputdir}{
	\code{numeric}, the chromosome index.
}
  \item{denoString}{
	\code{String}, column name for the pairwise comparison.	
}
  \item{time}{
	\code{String}, time generated by \code{Sys.time} function
}
  \item{toPlot}{
  \code{Boolean}, whether to plot lowess fit values
}
}
\details{
subfunction of \code{getResult}.
}

\author{
Sankar Martial
}
\seealso{
 See Also as \code{\link{getResult}}
}
\keyword{ ratio }