\name{quantileNorm}
\alias{quantileNorm}
\title{
quantile normalization}
\description{
\code{lowessNorm}, perform the quantile normalisation. Called by the \code{getNorm} function.
}
\usage{
quantileNorm(listInt, plotMVA = FALSE)
}
\arguments{
  \item{listInt}{
  	\code{numeric}, vector of intensity (SWI) values. 
}
  \item{plotMVA}{
  	\code{Boolean}, whether to display the MVA plot (on the X).
}
}
\details{
subfunction of \code{getNorm}.

}
\author{
Sankar Martial
}
\seealso{
See Also \code{\link{getNorm}}
}
\keyword{ normalization }
